requirements.txt
Django==4.2.7
boto3==1.29.3
pandas==2.1.3
Pillow==10.1.0
python-dotenv==1.0.0
